# internal/api/openapi.yml
openapi: 3.0.3
info:
  title: DBBat API
  description: |
    REST API for DBBat PostgreSQL Observability Proxy.

    DBBat is a transparent PostgreSQL proxy for query observability, access control, and safety.

    ## Authentication

    The API supports two authentication methods:

    - **Basic Auth**: HTTP Basic Authentication using username and password
    - **Bearer Token**: API key authentication using `Authorization: Bearer <api-key>`

    API keys cannot create or revoke other API keys (security restriction) - these operations require Basic Auth.

    ## Roles

    Users have one or more roles that determine their access:

    - **admin**: Full access to all resources and operations
    - **viewer**: Read-only access to observability data (connections, queries, audit)
    - **connector**: Can only access databases they have active grants for

    ## Password Change Requirement

    Users must change their initial password before logging in.
    Login attempts from users who haven't changed their password will receive a `403 Forbidden` response with error code `password_change_required`.
    Users must then call `PUT /auth/password` with their username and credentials to change their password before they can log in.

    ## Rate Limiting

    All authenticated endpoints are rate-limited. Response headers include:

    - `X-RateLimit-Limit`: Maximum requests per minute
    - `X-RateLimit-Remaining`: Remaining requests in current window
...

...

...

    - 1-2 failures: no delay
...

...
servers:
...
security:
  - basicAuth: []
  - bearerAuth: []
paths:
...
  /auth/login:
    post:
      tags:
        - Auth
      summary: Login and create web session
      description: |
        Authenticates with username/password and creates a short-lived web session token.

        The token should be stored and used as a Bearer token for subsequent API calls.
        Web session tokens expire after 1 hour.

        **Important**: Users who haven't changed their initial password will receive a 403 error.
        They must first change their password using `PUT /auth/password` before logging in.
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Password change required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PasswordChangeRequiredError'
        '429':
          $ref: '#/components/responses/AuthRateLimited'
  /auth/logout:
    post:
      tags:
        - Auth
      summary: Logout and revoke web session
      description: |
        Revokes the current web session token.

        After logout, the token will no longer be valid for authentication.
      operationId: logout
      responses:
        '204':
          description: Logout successful
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/me:
    get:
      tags:
        - Auth
      summary: Get current user and session info
      description: |
        Returns information about the currently authenticated user and their session.

        Use this endpoint to:
        - Validate that a session token is still valid
        - Check if password change is required
        - Get session expiration time
      operationId: getCurrentUser
      responses:
        '200':
          description: Current user and session info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/password:
    put:
      tags:
        - Auth
      summary: Change password (pre-login)
      description: |
        Allows users to change their password before logging in.
        Used for initial password change or password reset flows.

        Does not require authentication - validates username/current_password.
        Subject to rate limiting (same as login).

        After changing password, users can login using `POST /auth/login`.
      operationId: changePasswordPreLogin
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PreLoginPasswordChangeRequest'
      responses:
        '200':
          description: Password changed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Weak password
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    enum: [weak_password]
                  message:
                    type: string
                required:
                  - error
                  - message
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          $ref: '#/components/responses/AuthRateLimited'
  /users:
    post:
      tags:
        - Users
      summary: Create a new user
      description: Creates a new user account. Requires admin role.
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '200':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'

    get:
      tags:
        - Users
      summary: List users
      description: |
        Returns a list of users.

        - Admins see all users
        - Non-admins see only themselves
      operationId: listUsers
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
  /users/{uid}:
    parameters:
      - $ref: '#/components/parameters/UserUID'

    get:
      tags:
        - Users
      summary: Get user by UID
      description: Retrieves a specific user by their UID.
      operationId: getUser
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'

    put:
      tags:
        - Users
      summary: Update user
      description: |
        Updates a user account.

        - Non-admins can only update their own password
        - Non-admins cannot change roles
        - API keys cannot change passwords (requires Basic Auth)
      operationId: updateUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'

    delete:
      tags:
        - Users
      summary: Delete user
      description: |
        Deletes a user account. Requires admin role.

        Note: Cannot delete your own user account.
      operationId: deleteUser
      responses:
        '200':
          description: User deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
  /users/{uid}/password:
    parameters:
      - $ref: '#/components/parameters/UserUID'

    put:
      tags:
        - Users
      summary: Change password (authenticated)
      description: |
        Changes a user's password. Requires re-authentication via username/password in the request body.

        **Important**: This endpoint does NOT use Bearer token authentication.
        Users must provide their username and current password for re-authentication.

        - Users can only change their own password
        - Admin users can change any user's password (with their own admin credentials)
        - Subject to rate limiting (same as login)
      operationId: changePassword
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '200':
          description: Password changed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Weak password
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    enum: [weak_password]
                  message:
                    type: string
                required:
                  - error
                  - message
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Cannot change another user's password (non-admin)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/AuthRateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
  /databases:
    post:
      tags:
        - Databases
      summary: Create database configuration
      description: Creates a new database configuration. Requires admin role.
      operationId: createDatabase
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDatabaseRequest'
      responses:
        '200':
          description: Database created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Database'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'

    get:
      tags:
        - Databases
      summary: List databases
      description: |
        Returns a list of database configurations.

        Response varies by role:
        - **Admin**: Full details (host, port, database_name, username, ssl_mode)
        - **Viewer**: Limited info (uid, name, description)
        - **Connector**: Only databases they have active grants for (limited info)
      operationId: listDatabases
      responses:
        '200':
          description: List of databases
          content:
            application/json:
              schema:
                type: object
                properties:
                  databases:
                    type: array
                    items:
                      oneOf:
                        - $ref: '#/components/schemas/Database'
                        - $ref: '#/components/schemas/DatabaseLimited'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
  /databases/{uid}:
    parameters:
      - $ref: '#/components/parameters/DatabaseUID'

    get:
      tags:
        - Databases
      summary: Get database by UID
      description: |
        Retrieves a specific database configuration.

        Response varies by role:
        - **Admin**: Full details
        - **Viewer**: Limited info
        - **Connector**: Only if they have an active grant (limited info)
      operationId: getDatabase
      responses:
        '200':
          description: Database details
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/Database'
                  - $ref: '#/components/schemas/DatabaseLimited'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'

    put:
      tags:
        - Databases
      summary: Update database
      description: Updates a database configuration. Requires admin role.
      operationId: updateDatabase
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDatabaseRequest'
      responses:
        '200':
          description: Database updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'

    delete:
      tags:
        - Databases
      summary: Delete database
      description: Deletes a database configuration. Requires admin role.
      operationId: deleteDatabase
      responses:
        '200':
          description: Database deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
  /grants:
    post:
      tags:
        - Grants
      summary: Create access grant
      description: Creates a new access grant for a user to a database. Requires admin role.
      operationId: createGrant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGrantRequest'
      responses:
        '200':
          description: Grant created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessGrant'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'

    get:
      tags:
        - Grants
      summary: List grants
      description: |
        Returns a list of access grants.

        Connectors can only see their own grants.
      operationId: listGrants
      parameters:
        - name: user_id
          in: query
          description: Filter by user UID
          schema:
            type: string
            format: uuid
        - name: database_id
          in: query
          description: Filter by database UID
          schema:
            type: string
            format: uuid
        - name: active_only
          in: query
          description: Only return active (non-revoked, within time window) grants
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: List of grants
          content:
            application/json:
              schema:
                type: object
                properties:
                  grants:
                    type: array
                    items:
                      $ref: '#/components/schemas/AccessGrant'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
  /grants/{uid}:
    parameters:
      - $ref: '#/components/parameters/GrantUID'

    get:
      tags:
        - Grants
      summary: Get grant by UID
      description: |
        Retrieves a specific access grant.

        Connectors can only see their own grants.
      operationId: getGrant
      responses:
        '200':
          description: Grant details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessGrant'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'

    delete:
      tags:
        - Grants
      summary: Revoke grant
      description: Revokes an access grant. Requires admin role.
      operationId: revokeGrant
      responses:
        '200':
          description: Grant revoked successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /connections:
    get:
      tags:
        - Connections
      summary: List connections
      description: |
        Returns a list of proxy connections.

        Connectors can only see their own connections.
      operationId: listConnections
      parameters:
        - name: user_id
          in: query
          description: Filter by user UID
          schema:
            type: string
            format: uuid
        - name: database_id
          in: query
          description: Filter by database UID
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: List of connections
          content:
            application/json:
              schema:
                type: object
                properties:
                  connections:
                    type: array
                    items:
                      $ref: '#/components/schemas/Connection'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      description: HTTP Basic Authentication (username:password)
    bearerAuth:
      type: http
      scheme: bearer
      description: API key authentication (Bearer token)
...
  schemas:
...
    LoginRequest:
      type: object
      properties:
        username:
          type: string
          description: Username
        password:
          type: string
          format: password
          description: Password
      required:
        - username
        - password
...